requirements.txt
python-dotenv
streamlit
google-generativeai
streamlit-option-menu
yt_dlp
youtube-transcript-api